generate-random-str/requirements.txt
flask
redis
openai==0.28.0
python-dotenv